fix: item loading state, item update and delete, type in get orders by category request, move ItemDetails to ItemManagement page ui
