--- frontend/src/widgets/ItemDetails/index.ts
@@ -1,3 +1,0 @@
-import ItemDetails from "./ui/ItemDetails";
-
-export { ItemDetails };
