--- frontend/src/widgets/ItemDetails/ui/itemDetails.scss
@@ -1,101 +1,0 @@
-@import '../../../app/styles/mixins/global.scss';
-
-.itemDetails {
-  padding: 2rem;
-  font-size: 2rem;
-
-  &__form {
-    &-delete {
-      text-align: center;
-      @include hover;
-    }
-
-    &__section {
-      display: grid;
-      grid-template-columns: 1fr 2fr;
-      gap: 4rem;
-
-      @include max700 {
-        gap: 1rem;
-      }
-      @include max500 {
-        display: block;
-        gap: 2rem;
-      }
-      & textarea {
-        height: 8rem;
-        padding-left: 1rem;
-        font-family: inherit;
-      }
-    }
-    &__box {
-      &:not(:has(textarea)) {
-        @include input-box;
-      }
-      &:has(textarea),
-      &:has(input[type='file']) {
-        display: flex;
-        flex-direction: column;
-        gap: 0.5rem;
-        width: 100%;
-      }
-
-      &-select {
-        height: 4rem;
-        padding-left: 1rem;
-      }
-
-      // &__images {
-      //   margin: 4rem 0;
-      //   display: grid;
-      //   grid-template-columns: repeat(4, 1fr);
-      //   column-gap: 1rem;
-      //   & img {
-      //     width: 24rem;
-      //   }
-      //   &-img {
-      //     width: 24rem;
-      //     height: 24rem;
-      //     position: relative;
-
-      //     & p {
-      //       padding: 0.5rem;
-      //       background-color: $color-grey--light;
-
-      //       text-align: center;
-      //       width: 100%;
-      //       position: absolute;
-      //       bottom: 0;
-      //       left: 50%;
-      //       transform: translateX(-50%);
-
-      //       &:not(p.makeCover) {
-      //         background-color: $color-grey--medium;
-      //       }
-
-      //       &.makeCover {
-      //         cursor: pointer;
-      //         &:hover {
-      //           background-color: $color-grey--medium;
-      //         }
-      //       }
-      //     }
-      //   }
-      // }
-    }
-
-    &-buttons {
-      margin: 0 auto;
-      margin-top: 2rem;
-      @include center;
-      gap: 1rem;
-    }
-  }
-
-  @include max700 {
-    padding: 1rem 0rem;
-  }
-  @include max500 {
-    padding: 1rem;
-  }
-}
